refactor: char escapes
# Conflicts:
#	src/constructs/character-class.ts
